Add files via upload
Java project inspired from the game Four In A Row

Four in a row  is a two-player strategy game. Its goal is to be the player to first connect four stones of the same color in one horizontal, vertical or diagonal line. The classic game is played on a vertical and perforated board into which the players drop their stones one after the other. The board has seven columns and six rows. Each player gets twenty-one stones of the same color. The game ends in a draw when each cell is filled with a stone (board is full) and no "four in a row" is achieved.